+++ AGDevX.Spider.WebApi/Startup/Middleware.cs
@@ -1,8 +1,8 @@
 ﻿using AGDevX.Enums;
-using AGDevX.Environments;
 using AGDevX.Hosts;
 using AGDevX.Spider.WebApi.AuthZ;
 using AGDevX.Spider.WebApi.Config;
+using AGDevX.Spider.WebApi.Environment;
 using AGDevX.Web.Exceptions;
 using AGDevX.Web.Swagger;
 using Microsoft.AspNetCore.Builder;
@@ -16,7 +16,7 @@ public static class Middlware
     public static WebApplication ConfigureMiddlware(this WebApplication webApi, ApiConfig apiConfig)
     {
         if (apiConfig.Environment.IsOneOf(EnvironmentType.Local, EnvironmentType.Dev, EnvironmentType.Development)
-            || webApi.Environment.IsOneOf(EnvironmentType.Local, EnvironmentType.Dev)
+            || webApi.Environment.IsOneOf(EnvironmentType.Local.StringValue(), EnvironmentType.Dev.StringValue())
             || webApi.Environment.IsDevelopment())
         {
         }
