+++ AGDevX/Hosts/IHostEnvironmentExtensions.cs
@@ -1,20 +1,17 @@
-﻿using System.Linq;
-using AGDevX.Enums;
-using AGDevX.Environments;
-using AGDevX.IEnumerables;
+﻿using AGDevX.IEnumerables;
 using Microsoft.Extensions.Hosting;
 
 namespace AGDevX.Hosts;
 
 public static class IHostEnvironmentExtensions
 {
-    public static bool IsOneOf(this IHostEnvironment? webHostEnvironment, params EnvironmentType[]? environments)
+    public static bool IsOneOf(this IHostEnvironment? webHostEnvironment, params string[]? environments)
     {
         if (webHostEnvironment == null || environments.IsNullOrEmpty())
         {
             return false;
         }
 
-        return environments!.Select(env => env.StringValue()).ContainsIgnoreCase(webHostEnvironment.EnvironmentName);
+        return environments!.ContainsIgnoreCase(webHostEnvironment.EnvironmentName);
     }
 }
